// src/public-api/model/bannerDto.ts
/**
 * MISCore Shop API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BannerDto { 
    id?: number;
    title: string;
    description?: string | null;
    imageUrl: string;
    linkUrl?: string | null;
    displayOrder?: number;
    isActive?: boolean;
    createdAt?: string;
    updatedAt?: string | null;
    bannerType?: string | null;
    startDate?: string | null;
    endDate?: string | null;
}